devlib/trace: Refactor `trace` to be `collector`
We now have have multiple `trace` instruments that do not match the
description and therefore are moved to have a more suitably named
hierarchy.
